Don't specify `to` when making the request to mailserver.
In some cases (say the clock of the device has drifted) specifying a `to` parameter
is decrimental and might result in missing messages.

We have changed the mailservers to default `to` if not present to `now`,
which should gives us better guarantees.

We also removed the 24 hours limit as now all the requests will be
paginated according to the `limit` parameter.